# Further parallelization of the JPlag pipeline