--- a/languages/c/src/main/java/de/jplag/c/CLanguage.java
+++ b/languages/c/src/main/java/de/jplag/c/CLanguage.java
@@ -15,12 +15,6 @@ public class CLanguage implements Language {
     private static final String NAME = "C";
     private static final String IDENTIFIER = "c";
 
-    private final Scanner scanner; // c code is scanned not parsed
-
-    public CLanguage() {
-        scanner = new Scanner();
-    }
-
     @Override
     public String[] suffixes() {
         return new String[] {".cpp", ".CPP", ".cxx", ".CXX", ".c++", ".C++", ".c", ".C", ".cc", ".CC", ".h", ".H", ".hpp", ".HPP", ".hh", ".HH"};
@@ -43,6 +37,6 @@ public class CLanguage implements Language {
 
     @Override
     public List<Token> parse(Set<File> files, boolean normalize) throws ParsingException {
-        return this.scanner.scan(files);
+        return new Scanner().scan(files);
     }
 }