--- a/languages/cpp/src/main/java/de/jplag/cpp/CPPLanguage.java
+++ b/languages/cpp/src/main/java/de/jplag/cpp/CPPLanguage.java
@@ -1,21 +1,22 @@
 package de.jplag.cpp;
 
+import java.io.File;
+import java.util.List;
+import java.util.Set;
+
 import org.kohsuke.MetaInfServices;
 
 import de.jplag.Language;
-import de.jplag.antlr.AbstractAntlrLanguage;
+import de.jplag.ParsingException;
+import de.jplag.Token;
 
 /**
  * The entry point for the ANTLR parser based C++ language module.
  */
 @MetaInfServices(Language.class)
-public class CPPLanguage extends AbstractAntlrLanguage {
+public class CPPLanguage implements Language {
     private static final String NAME = "C++";
     private static final String IDENTIFIER = "cpp";
-
-    public CPPLanguage() {
-        super(new CPPParserAdapter());
-    }
 
     @Override
     public String[] suffixes() {
@@ -45,5 +46,10 @@ public class CPPLanguage extends AbstractAntlrLanguage {
     @Override
     public boolean supportsNormalization() {
         return true;
+    }
+
+    @Override
+    public List<Token> parse(Set<File> files, boolean normalize) throws ParsingException {
+        return new CPPParserAdapter().parse(files);
     }
 }