--- a/languages/csharp/src/main/java/de/jplag/csharp/CSharpLanguage.java
+++ b/languages/csharp/src/main/java/de/jplag/csharp/CSharpLanguage.java
@@ -1,22 +1,24 @@
 package de.jplag.csharp;
 
+import java.io.File;
+import java.util.List;
+import java.util.Set;
+
 import org.kohsuke.MetaInfServices;
 
-import de.jplag.antlr.AbstractAntlrLanguage;
+import de.jplag.Language;
+import de.jplag.ParsingException;
+import de.jplag.Token;
 
 /**
  * C# language with full support of C# 6 features and below.
  */
 @MetaInfServices(de.jplag.Language.class)
-public class CSharpLanguage extends AbstractAntlrLanguage {
+public class CSharpLanguage implements Language {
     private static final String NAME = "C#";
     private static final String IDENTIFIER = "csharp";
     private static final String[] FILE_ENDINGS = new String[] {".cs", ".CS"};
     private static final int DEFAULT_MIN_TOKEN_MATCH = 8;
-
-    public CSharpLanguage() {
-        super(new CSharpParserAdapter());
-    }
 
     @Override
     public String[] suffixes() {
@@ -36,5 +38,10 @@ public class CSharpLanguage extends AbstractAntlrLanguage {
     @Override
     public int minimumTokenMatch() {
         return DEFAULT_MIN_TOKEN_MATCH;
+    }
+
+    @Override
+    public List<Token> parse(Set<File> files, boolean normalize) throws ParsingException {
+        return new CSharpParserAdapter().parse(files);
     }
 }