--- a/languages/emf-model/src/main/java/de/jplag/emf/model/EmfModelLanguage.java
+++ b/languages/emf-model/src/main/java/de/jplag/emf/model/EmfModelLanguage.java
@@ -3,9 +3,12 @@ package de.jplag.emf.model;
 import java.io.File;
 import java.util.Comparator;
 import java.util.List;
+import java.util.Set;
 
 import org.kohsuke.MetaInfServices;
 
+import de.jplag.ParsingException;
+import de.jplag.Token;
 import de.jplag.emf.dynamic.DynamicEmfLanguage;
 import de.jplag.emf.model.parser.DynamicModelParser;
 
@@ -20,10 +23,6 @@ public class EmfModelLanguage extends DynamicEmfLanguage {
     private static final String IDENTIFIER = "emf-model";
 
     public static final String VIEW_FILE_SUFFIX = ".treeview";
-
-    public EmfModelLanguage() {
-        super(new DynamicModelParser());
-    }
 
     @Override
     public String[] suffixes() {
@@ -53,5 +52,10 @@ public class EmfModelLanguage extends DynamicEmfLanguage {
     @Override
     public List<File> customizeSubmissionOrder(List<File> sub) {
         return sub.stream().sorted(Comparator.comparing(file -> file.getName().endsWith(FILE_ENDING) ? 0 : 1)).toList();
+    }
+
+    @Override
+    public List<Token> parse(Set<File> files, boolean normalize) throws ParsingException {
+        return new DynamicModelParser().parse(files, normalize);
     }
 }