--- a/languages/golang/src/main/java/de/jplag/golang/GoLanguage.java
+++ b/languages/golang/src/main/java/de/jplag/golang/GoLanguage.java
@@ -1,19 +1,21 @@
 package de.jplag.golang;
 
+import java.io.File;
+import java.util.List;
+import java.util.Set;
+
 import org.kohsuke.MetaInfServices;
 
-import de.jplag.antlr.AbstractAntlrLanguage;
+import de.jplag.Language;
+import de.jplag.ParsingException;
+import de.jplag.Token;
 
 @MetaInfServices(de.jplag.Language.class)
-public class GoLanguage extends AbstractAntlrLanguage {
+public class GoLanguage implements Language {
     private static final String NAME = "Go";
     private static final String IDENTIFIER = "go";
     private static final int DEFAULT_MIN_TOKEN_MATCH = 8;
     private static final String[] FILE_EXTENSIONS = {".go"};
-
-    public GoLanguage() {
-        super(new GoParserAdapter());
-    }
 
     @Override
     public String[] suffixes() {
@@ -33,5 +35,10 @@ public class GoLanguage extends AbstractAntlrLanguage {
     @Override
     public int minimumTokenMatch() {
         return DEFAULT_MIN_TOKEN_MATCH;
+    }
+
+    @Override
+    public List<Token> parse(Set<File> files, boolean normalize) throws ParsingException {
+        return new GoParserAdapter().parse(files);
     }
 }