--- a/languages/kotlin/src/main/java/de/jplag/kotlin/KotlinLanguage.java
+++ b/languages/kotlin/src/main/java/de/jplag/kotlin/KotlinLanguage.java
@@ -1,23 +1,25 @@
 package de.jplag.kotlin;
 
+import java.io.File;
+import java.util.List;
+import java.util.Set;
+
 import org.kohsuke.MetaInfServices;
 
-import de.jplag.antlr.AbstractAntlrLanguage;
+import de.jplag.Language;
+import de.jplag.ParsingException;
+import de.jplag.Token;
 
 /**
  * This represents the Kotlin language as a language supported by JPlag.
  */
 @MetaInfServices(de.jplag.Language.class)
-public class KotlinLanguage extends AbstractAntlrLanguage {
+public class KotlinLanguage implements Language {
 
     private static final String NAME = "Kotlin";
     private static final String IDENTIFIER = "kotlin";
     private static final int DEFAULT_MIN_TOKEN_MATCH = 8;
     private static final String[] FILE_EXTENSIONS = {".kt"};
-
-    public KotlinLanguage() {
-        super(new KotlinParserAdapter());
-    }
 
     @Override
     public String[] suffixes() {
@@ -37,5 +39,10 @@ public class KotlinLanguage extends AbstractAntlrLanguage {
     @Override
     public int minimumTokenMatch() {
         return DEFAULT_MIN_TOKEN_MATCH;
+    }
+
+    @Override
+    public List<Token> parse(Set<File> files, boolean normalize) throws ParsingException {
+        return new KotlinParserAdapter().parse(files);
     }
 }