--- a/languages/python-3/src/main/java/de/jplag/python3/PythonLanguage.java
+++ b/languages/python-3/src/main/java/de/jplag/python3/PythonLanguage.java
@@ -1,17 +1,19 @@
 package de.jplag.python3;
 
+import java.io.File;
+import java.util.List;
+import java.util.Set;
+
 import org.kohsuke.MetaInfServices;
 
-import de.jplag.antlr.AbstractAntlrLanguage;
+import de.jplag.Language;
+import de.jplag.ParsingException;
+import de.jplag.Token;
 
 @MetaInfServices(de.jplag.Language.class)
-public class PythonLanguage extends AbstractAntlrLanguage {
+public class PythonLanguage implements Language {
     private static final String NAME = "Python";
     private static final String IDENTIFIER = "python3";
-
-    public PythonLanguage() {
-        super(new PythonParserAdapter());
-    }
 
     @Override
     public String[] suffixes() {
@@ -31,5 +33,10 @@ public class PythonLanguage extends AbstractAntlrLanguage {
     @Override
     public int minimumTokenMatch() {
         return 12;
+    }
+
+    @Override
+    public List<Token> parse(Set<File> files, boolean normalize) throws ParsingException {
+        return new PythonParserAdapter().parse(files);
     }
 }