--- a/languages/scala/src/main/scala/de/jplag/scala/ScalaLanguage.scala
+++ b/languages/scala/src/main/scala/de/jplag/scala/ScalaLanguage.scala
@@ -9,7 +9,6 @@ import java.util
 import scala.jdk.CollectionConverters.{SeqHasAsJava, SetHasAsScala}
 
 class ScalaLanguage extends de.jplag.Language {
-  private val parser = new Parser
   private final val fileExtensions = Array(".scala", ".sc")
 
   override def suffixes: Array[String] = fileExtensions
@@ -20,5 +19,5 @@ class ScalaLanguage extends de.jplag.Language {
 
   override def minimumTokenMatch = 8
 
-  override def parse(files: util.Set[File], normalize: Boolean): java.util.List[Token] = this.parser.parse(files.asScala.toSet).asJava
+  override def parse(files: util.Set[File], normalize: Boolean): java.util.List[Token] = new Parser().parse(files.asScala.toSet).asJava
 }