--- a/languages/swift/src/main/java/de/jplag/swift/SwiftLanguage.java
+++ b/languages/swift/src/main/java/de/jplag/swift/SwiftLanguage.java
@@ -1,24 +1,26 @@
 package de.jplag.swift;
 
+import java.io.File;
+import java.util.List;
+import java.util.Set;
+
 import org.kohsuke.MetaInfServices;
 
-import de.jplag.antlr.AbstractAntlrLanguage;
+import de.jplag.Language;
+import de.jplag.ParsingException;
+import de.jplag.Token;
 
 /**
  * This represents the Swift language as a language supported by JPlag.
  */
 @MetaInfServices(de.jplag.Language.class)
-public class SwiftLanguage extends AbstractAntlrLanguage {
+public class SwiftLanguage implements Language {
 
     private static final String IDENTIFIER = "swift";
 
     private static final String NAME = "Swift";
     private static final int DEFAULT_MIN_TOKEN_MATCH = 8;
     private static final String[] FILE_EXTENSIONS = {".swift"};
-
-    public SwiftLanguage() {
-        super(new SwiftParserAdapter());
-    }
 
     @Override
     public String[] suffixes() {
@@ -38,5 +40,10 @@ public class SwiftLanguage extends AbstractAntlrLanguage {
     @Override
     public int minimumTokenMatch() {
         return DEFAULT_MIN_TOKEN_MATCH;
+    }
+
+    @Override
+    public List<Token> parse(Set<File> files, boolean normalize) throws ParsingException {
+        return new SwiftParserAdapter().parse(files);
     }
 }