--- a/languages/typescript/src/main/java/de/jplag/typescript/TypeScriptLanguage.java
+++ b/languages/typescript/src/main/java/de/jplag/typescript/TypeScriptLanguage.java
@@ -1,14 +1,20 @@
 package de.jplag.typescript;
 
+import java.io.File;
+import java.util.List;
+import java.util.Set;
+
 import org.kohsuke.MetaInfServices;
 
-import de.jplag.antlr.AbstractAntlrLanguage;
+import de.jplag.Language;
+import de.jplag.ParsingException;
+import de.jplag.Token;
 
 /**
  * This represents the TypeScript language as a language supported by JPlag.
  */
 @MetaInfServices(de.jplag.Language.class)
-public class TypeScriptLanguage extends AbstractAntlrLanguage {
+public class TypeScriptLanguage implements Language {
 
     private static final String IDENTIFIER = "typescript";
     private static final String NAME = "TypeScript";
@@ -40,7 +46,7 @@ public class TypeScriptLanguage extends AbstractAntlrLanguage {
     }
 
     @Override
-    protected TypeScriptParserAdapter initializeParser(boolean normalize) {
-        return new TypeScriptParserAdapter(getOptions().useStrictDefault());
+    public List<Token> parse(Set<File> files, boolean normalize) throws ParsingException {
+        return new TypeScriptParserAdapter(options.useStrictDefault()).parse(files);
     }
 }