… language-antlr-utils/src/main/java/de/jplag/antlr/AbstractAntlrLanguage.java
… language-antlr-utils/src/test/java/de/jplag/antlr/testLanguage/TestLanguage.java