Introduce parallelization of the submission parsing and the token sequence normalization.

For a dataset of 868 Java programs with a total of 1.3 million LOC, we observe the following speedup on a M1 Pro MacBook:

- Parsing: Reduced from 25.7 to 8.3 seconds
- Normalization: Reduced from 14.3 to 4.8 seconds

Left TODO:

- [ ] check if the logging and 
- [x] Parallel normalization
- [x] Parallel parsing
- [x] Sequential parsing for EMF modules
- [x] Avoid race conditions for error counting and current submission name storing.